+++ package/mecab/CMakeLists.txt
@@ -1,9 +1,22 @@
 cmake_minimum_required(VERSION 2.6)
 find_library(HAVE_LIBMECAB NAMES mecab)
+
 if(HAVE_LIBMECAB)
-	set(ExtraLibs ${ExtraLibs} ${HAVE_LIBMECAB})
-	set(PACKAGE_EXTRA_LIBRARY ${ExtraLibs})
-	set(PACKAGE_SOURCE_CODE mecab_glue.c)
-	set(PACKAGE_SCRIPT_CODE mecab_glue.k)
-	add_konoha_package(mecab)
+	if(APPLE)
+		string(COMPARE EQUAL "/usr/lib/libmecab.dylib" ${HAVE_LIBMECAB} USE_ORIGINAL_LIB)
+		if(${USE_ORIGINAL_LIB})
+		else(${USE_ORIGINAL_LIB})
+			set(ExtraLibs ${ExtraLibs} ${HAVE_LIBMECAB})
+			set(PACKAGE_EXTRA_LIBRARY ${ExtraLibs})
+			set(PACKAGE_SOURCE_CODE mecab_glue.c)
+			set(PACKAGE_SCRIPT_CODE mecab_glue.k)
+			add_konoha_package(mecab)
+		endif(${USE_ORIGINAL_LIB})
+	else(APPLE)
+		set(ExtraLibs ${ExtraLibs} ${HAVE_LIBMECAB})
+		set(PACKAGE_EXTRA_LIBRARY ${ExtraLibs})
+		set(PACKAGE_SOURCE_CODE mecab_glue.c)
+		set(PACKAGE_SCRIPT_CODE mecab_glue.k)
+		add_konoha_package(mecab)
+	endif(APPLE)
 endif(HAVE_LIBMECAB)
